Moved remaining functions to renderer and interactivity files
